+++ src/Illuminate/View/Engines/PhpEngine.php
@@ -4,9 +4,10 @@ namespace Illuminate\View\Engines;
 
 use Exception;
 use Throwable;
+use Illuminate\Contracts\View\Engine;
 use Symfony\Component\Debug\Exception\FatalThrowableError;
 
-class PhpEngine implements EngineInterface
+class PhpEngine implements Engine
 {
     /**
      * Get the evaluated contents of the view.
